+++ src/components/NavigationBar/NavigationBar.tsx
@@ -9,36 +9,13 @@ import Menu from '@mui/material/Menu';
 import Container from '@mui/material/Container';
 import Button from '@mui/material/Button';
 import MenuItem from '@mui/material/MenuItem';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-
+import Theme from 'styles/themes/Theme';
 import { ReactComponent as Paw } from 'assets/paw-solid.svg';
 import { makeStyles } from '@mui/styles';
 import imgLogo from 'assets/logo.png';
 import 'styles/fonts.scss';
 
 const pages = ['About Us', 'Contact', 'Sign Up', 'Log In'];
-
-const theme = createTheme({
-  components: {
-    MuiAppBar: {
-      styleOverrides: {
-        root: {
-          backgroundColor: '#16bac6',
-        },
-      },
-    },
-    MuiButton: {
-      styleOverrides: {
-        root: {        
-          fontFamily: '"Ubuntu", sans-serif',
-          textTransform: 'capitalize',
-          color: '#112025',
-          fontSize: '1rem'
-        },
-      },
-    },
-  },
-});
 
 const useStyles = makeStyles({
   root: {
@@ -78,7 +55,7 @@ export const NavigationBar = () => {
   const classes = useStyles();
 
   return (
-    <ThemeProvider theme={theme}>
+    <Theme >
       <AppBar position="static">
         <Container maxWidth="xl" className={classes.root}>
           <Toolbar disableGutters>
@@ -122,8 +99,7 @@ export const NavigationBar = () => {
               >
                 {pages.map((page) => (
                   <MenuItem key={page} onClick={handleCloseNavMenu}>
-                    <Typography textAlign="center">{page}
-                    </Typography>
+                    <Typography textAlign="center">{page}</Typography>
                   </MenuItem>
                 ))}
               </Menu>
@@ -151,6 +127,6 @@ export const NavigationBar = () => {
           </Toolbar>
         </Container>
       </AppBar>
-    </ThemeProvider>
+    </Theme>
   );
 };
