+++ src/components/inputs/Input.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { makeStyles } from '@material-ui/styles';
+import { makeStyles } from '@mui/styles';
 import TextField from '@mui/material/TextField';
